API JWT Authentication

+++ controller/api/v1/posts_api.js
@@ -21,17 +21,19 @@ module.exports.index = async function(req, res){
 module.exports.destroy =async function(req,res){
     try{
         const post =await Post.findById(req.params.id)
-        // if(post.user == req.user.id){
+        if(post.user == req.user.id){
             post.deleteOne();
             await Comment.deleteMany({post:req.params.id});
 
             return res.status(200).json({
                 message : "post & associated comment deleted successfully"
             })
-        // }
-        // else{
-        //     return res.redirect('back');
-        // }
+        }
+        else{
+            return res.status(401).json({
+                message : "you cannot delete this post"
+            })
+        }
     }catch(err){
         console.log("****",err)
         return res.status(500).json({
